+++ coralogix/resource_coralogix_dashboard_test.go
@@ -31,6 +31,10 @@ func TestAccCoralogixResourceDashboard(t *testing.T) {
 					resource.TestCheckResourceAttrSet(dashboardResourceName, "id"),
 					resource.TestCheckResourceAttr(dashboardResourceName, "name", "test"),
 					resource.TestCheckResourceAttr(dashboardResourceName, "description", "dashboards team is messing with this 🗿"),
+					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.options.name", "Status"),
+					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.options.color", "blue"),
+					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.options.description", "abc"),
+					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.options.collapsed", "false"),
 					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.rows.0.height", "19"),
 					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.rows.0.widgets.0.title", "status 4XX"),
 					resource.TestCheckResourceAttr(dashboardResourceName, "layout.sections.0.rows.0.widgets.0.definition.line_chart.query_definitions.0.query.metrics.promql_query", "http_requests_total{status!~\"4..\"}"),
@@ -145,6 +149,12 @@ func testAccCoralogixResourceDashboard() string {
   layout      = {
     sections = [
       {
+        options = {
+          name = "Status"
+          description = "abc"
+          collapsed = false
+          color = "blue"
+        }
         rows = [
           {
             height  = 19