+++ coralogix/resource_coralogix_sli.go
@@ -125,7 +125,8 @@ func (r *SLIResource) ImportState(ctx context.Context, req resource.ImportStateR
 
 func (r *SLIResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
-		Version: 0,
+		DeprecationMessage: "Please upgrade to using the 'coralogix_slo' resource",
+		Version:            0,
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
 				Computed: true,